refactor: make code look better and add caching
